backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.9.2
agno==1.3.4
google-generativeai==0.8.3
google-genai>=1.0.0
gitpython==3.1.44
docker==7.1.0
aiofiles==24.1.0
httpx==0.28.0
python-multipart==0.0.20
